feat(usecolorscheme): some refactoring of dark mode
